# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MovieSearchResult {
  tmdbId: number;
  titlePtBr: string;
  originalTitle: string;
  /** @nullable */
  year: number | null;
  /** @nullable */
  rating: number | null;
  /** @nullable */
  posterUrl: string | null;
}

export interface MovieDetail {
  tmdbId: number;
  titlePtBr: string;
  originalTitle: string;
  /** @nullable */
  year: number | null;
  /** @nullable */
  rating: number | null;
  /** @nullable */
  posterUrl: string | null;
  /** @nullable */
  backdropUrl: string | null;
  /**
   * Synopsis in Brazilian Portuguese (with English fallback)
   * @nullable
   */
  overview: string | null;
  /** @nullable */
  runtime: number | null;
  genres: string[];
  /** @nullable */
  tagline: string | null;
}

export interface ErrorResponse {
  error: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export type MovieStatus = (typeof MovieStatus)[keyof typeof MovieStatus];

export const MovieStatus = {
  want: "want",
  watched: "watched",
  abandoned: "abandoned",
} as const;

export interface UserMovie {
  id: string;
  /** @nullable */
  tmdbId: number | null;
  titlePtBr: string;
  originalTitle: string;
  /** @nullable */
  year: number | null;
  /** @nullable */
  rating: number | null;
  /** @nullable */
  posterUrl: string | null;
  status: MovieStatus;
  addedAt: string;
}

export interface AddUserMovieRequest {
  /** @nullable */
  tmdbId?: number | null;
  /** @minLength 1 */
  titlePtBr: string;
  originalTitle: string;
  /** @nullable */
  year?: number | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  posterUrl?: string | null;
  status: MovieStatus;
}

export interface UpdateUserMovieRequest {
  status: MovieStatus;
}

export interface BulkUpsertRequest {
  movies: AddUserMovieRequest[];
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type SearchMoviesParams = {
  /**
   * @minLength 1
   */
  q: string;
};

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};