# lib/api-zod/src/generated/types/bulkUpsertRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AddUserMovieRequest } from "./addUserMovieRequest";

export interface BulkUpsertRequest {
  movies: AddUserMovieRequest[];
}