# lib/api-zod/src/generated/types/movieDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MovieDetail {
  tmdbId: number;
  titlePtBr: string;
  originalTitle: string;
  /** @nullable */
  year: number | null;
  /** @nullable */
  rating: number | null;
  /** @nullable */
  posterUrl: string | null;
  /** @nullable */
  backdropUrl: string | null;
  /**
   * Synopsis in Brazilian Portuguese (with English fallback)
   * @nullable
   */
  overview: string | null;
  /** @nullable */
  runtime: number | null;
  genres: string[];
  /** @nullable */
  tagline: string | null;
}