# lib/api-zod/src/generated/types/movieStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type MovieStatus = (typeof MovieStatus)[keyof typeof MovieStatus];

export const MovieStatus = {
  want: "want",
  watched: "watched",
  abandoned: "abandoned",
} as const;