# lib/api-zod/src/generated/types/userMovie.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MovieStatus } from "./movieStatus";

export interface UserMovie {
  id: string;
  /** @nullable */
  tmdbId: number | null;
  titlePtBr: string;
  originalTitle: string;
  /** @nullable */
  year: number | null;
  /** @nullable */
  rating: number | null;
  /** @nullable */
  posterUrl: string | null;
  status: MovieStatus;
  addedAt: Date;
}